requirements.txt
azure
azure-storage
O365
pymiscutils
pathmagic
pyiotools
pysubtypes
maybe-else
